Finished Option tests.
